requirements.txt
python-telegram-bot==22.7
supabase==2.15.0
APScheduler==3.10.4
python-dotenv==1.2.2
pytz==2026.1.post1
